fix(mcp): bind the MCP listener through the shared exclusive bind
The MCP HTTP transport bound with a plain tokio bind, leaving the port
open on Windows to the same same-user SO_REUSEADDR takeover the
DynamoDB listener was just protected against. The bind helpers move to
a shared net module so both servers use one path; Unix behaviour is
unchanged.
